+++ autoPyTorch/pipeline/components/setup/network_backbone/MLPBackbone.py
@@ -10,12 +10,8 @@ from ConfigSpace.hyperparameters import (
 
 from torch import nn
 
-from autoPyTorch.pipeline.components.setup.network_backbone.base_network_backbone import (
-    NetworkBackboneComponent,
-)
-from autoPyTorch.pipeline.components.setup.network_backbone.utils import (
-    _activations,
-)
+from autoPyTorch.pipeline.components.setup.network_backbone.base_network_backbone import NetworkBackboneComponent
+from autoPyTorch.pipeline.components.setup.network_backbone.utils import _activations
 
 
 class MLPBackbone(NetworkBackboneComponent):
@@ -28,7 +24,6 @@ class MLPBackbone(NetworkBackboneComponent):
         - Using or not dropout
         - Specifying the number of units per layers
     """
-    supported_tasks = {"tabular_classification", "tabular_regression"}
 
     def build_backbone(self, input_shape: Tuple[int, ...]) -> nn.Module:
         layers = list()  # type: List[nn.Module]
