+++ autoPyTorch/pipeline/components/setup/network_head/fully_connected.py
@@ -8,25 +8,15 @@ import numpy as np
 
 from torch import nn
 
-from autoPyTorch.pipeline.components.setup.network_head.base_network_head import (
-    NetworkHeadComponent,
-)
-
-_activations: Dict[str, nn.Module] = {
-    "relu": nn.ReLU,
-    "tanh": nn.Tanh,
-    "sigmoid": nn.Sigmoid
-}
+from autoPyTorch.pipeline.components.setup.network_head.base_network_head import NetworkHeadComponent
+from autoPyTorch.pipeline.components.setup.network_head.utils import _activations
 
 
 class FullyConnectedHead(NetworkHeadComponent):
     """
-    Standard head consisting of a number of fully connected layers.
+    Head consisting of a number of fully connected layers.
     Flattens any input in a array of shape [B, prod(input_shape)].
     """
-    supported_tasks = {"tabular_classification", "tabular_regression",
-                       "image_classification", "image_regression",
-                       "time_series_classification", "time_series_regression"}
 
     def build_head(self, input_shape: Tuple[int, ...], output_shape: Tuple[int, ...]) -> nn.Module:
         layers = [nn.Flatten()]
@@ -47,8 +37,8 @@ class FullyConnectedHead(NetworkHeadComponent):
             'shortname': 'FullyConnectedHead',
             'name': 'FullyConnectedHead',
             'handles_tabular': True,
-            'handles_image': False,
-            'handles_time_series': False,
+            'handles_image': True,
+            'handles_time_series': True,
         }
 
     @staticmethod
