+++ autoPyTorch/pipeline/components/setup/network_head/fully_convolutional.py
@@ -7,15 +7,8 @@ from ConfigSpace.hyperparameters import CategoricalHyperparameter, UniformIntege
 import torch
 from torch import nn
 
-from autoPyTorch.pipeline.components.setup.network_head.base_network_head import (
-    NetworkHeadComponent,
-)
-
-_activations: Dict[str, nn.Module] = {
-    "relu": nn.ReLU,
-    "tanh": nn.Tanh,
-    "sigmoid": nn.Sigmoid
-}
+from autoPyTorch.pipeline.components.setup.network_head.base_network_head import NetworkHeadComponent
+from autoPyTorch.pipeline.components.setup.network_head.utils import _activations
 
 
 class _FullyConvolutional2DHead(nn.Module):
@@ -56,7 +49,6 @@ class FullyConvolutional2DHead(NetworkHeadComponent):
     Head consisting of a number of 2d convolutional connected layers.
     Applies a global pooling operation in the end.
     """
-    supported_tasks = {"image_classification", "image_regression"}
 
     def build_head(self, input_shape: Tuple[int, ...], output_shape: Tuple[int, ...]) -> nn.Module:
         return _FullyConvolutional2DHead(input_shape=input_shape,
@@ -70,8 +62,8 @@ class FullyConvolutional2DHead(NetworkHeadComponent):
     @staticmethod
     def get_properties(dataset_properties: Optional[Dict[str, Any]] = None) -> Dict[str, Union[str, bool]]:
         return {
-            'shortname': 'FullyConvolutionalHead',
-            'name': 'FullyConvolutionalHead',
+            'shortname': 'FullyConvolutional2DHead',
+            'name': 'FullyConvolutional2DHead',
             'handles_tabular': False,
             'handles_image': True,
             'handles_time_series': False,
