+++ autoPyTorch/pipeline/components/setup/network_head/utils.py
@@ -1,0 +1,7 @@
+import torch
+
+_activations = {
+    "relu": torch.nn.ReLU,
+    "tanh": torch.nn.Tanh,
+    "sigmoid": torch.nn.Sigmoid
+}
